Add files via upload
The purpose of the Bus Reservation System is to increase the time efficiently in the workplaces and that can be accomplished through the use of our reservation system. Employees will be allowed to perform longer and more demanding shifts to achieve a certain rate or income, which will benefit the company they work for. This system will guarantee our system is simple to use and understand for employees and companies, guaranteeing that all employees commute safely to and from their workplaces. It will make certain all employees are picked up on time and also guarantee that all staff arrive on schedule at their destination. This Bus Reservation System aims to ensure safe traveling of all employees as well as ensuring that they all make it to their destination within time. We will ensure that all customers are benefited; the individual customers will arrive at their workplace on schedule and travel in a secure way. In addition, the companies which utilize our service will not be short staff due to the lack of transportation